LPD-35336 Simplifying
We now use the same components used by batch engine. This means that for objects, the _objectDefinition is already set and we do not need to handle that. We just need to call the setCompany as batch engine does.

I've also simplified the company id injection and we will just be getting it from the thread local. This is because modifying the OpenAPIResource interface is risky as it could cause breaking changes with previous versions of REST Builder used by customers.

With this commit we reduce the scope of the change just to the vulcan infrastructure, making it transparent to the other modules except for objects in which we do need to set certain properties when registering the components.

Co-authored-by: Alejandro Tardín <alejandro.tardin@liferay.com>

# breaking

## What modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/resource/EntityModelResource.java

New method `public void setContextCompany(Company contextCompany)` to inject the context company of the request.

## Why

It is necessary to be able to get the EntityFields based on the company, so we need to be able to set the company context. We are following the existing pattern in other intefaces (such as VulcanBatchEngineTaskItemDelegate or <SchemaName>Resource). The method is always implemented by REST Builder, so it should not be considered a real breaking change unless the interface is manually implemented without implementing any <SchemaName>Resource interface.

----