LPD-61989 Fix test flakiness in ObjectEntry keywords due to async increment processing
Keyword/tag usage is updated asynchronously, which can cause test failures
if assertions run before the updates are flushed.

An extra POST is used to trigger a transaction commit, which helps flush
previous increments before the final assertions — in order to avoid a direct
Thread.sleep that also fixes the issue.